Release

+++ src/models/image.go
@@ -32,10 +32,10 @@ func (t ImageType) String() string {
 type Image struct {
 	bun.BaseModel `bun:"table:image"`
 
-	ID           uint      `bun:"id,pk,nullzero"`
-	OwnerID      uint      `bun:"owner_id,notnull"`
-	EntityID     uint      `bun:"entity_id,notnull"`
-	EntityType   uint      `bun:"entity_type,notnull"`
+	ID           uint64    `bun:"id,pk,nullzero"`
+	OwnerID      uint64    `bun:"owner_id,notnull"`
+	EntityID     uint64    `bun:"entity_id,notnull"`
+	EntityType   uint64    `bun:"entity_type,notnull"`
 	FileName     string    `bun:"file_name,notnull"`
 	OriginalName string    `bun:"original_name,notnull"`
 	Extension    string    `bun:"extension,notnull"`
